refactor: address tab code refactoring
